Réingénierie test_parsing. 	modifié:         tests/test_parsing.py Réajustement des dummydata pour coller avec test_parsing
